Updated domain module with interfaces
