+++ include/domain/interface.hpp
@@ -1,0 +1,7 @@
+#ifndef _DOMAIN_INTERFACE_HPP
+#define _DOMAIN_INTERFACE_HPP
+
+#include "domain.hpp"
+#include "elastic_domain.hpp"
+
+#endif
