+++ tests/unittests/domain/rmass_inverse_tests.cpp
@@ -1,14 +1,14 @@
-#include "../../../include/compute.h"
-#include "../../../include/domain.h"
-#include "../../../include/material.h"
-#include "../../../include/mesh.h"
-#include "../../../include/parameter_parser.h"
-#include "../../../include/quadrature.h"
-#include "../../../include/read_sources.h"
-#include "../../../include/utils.h"
 #include "../Kokkos_Environment.hpp"
 #include "../MPI_environment.hpp"
 #include "../utilities/include/compare_array.h"
+#include "compute.h"
+#include "domain/interface.hpp"
+#include "material.h"
+#include "mesh.h"
+#include "parameter_parser.h"
+#include "quadrature.h"
+#include "read_sources.h"
+#include "utils.h"
 #include "yaml-cpp/yaml.h"
 
 // ----- Parse test config ------------- //
